SAM3: force float32 model + autocast for label accuracy parity
float16 autocast reduced labels from ~155 to 11 on Sand2. Using
model.float() + torch.autocast(float32) to match HPC precision.

+++ pipeline/sam3_segment.py
@@ -234,7 +234,7 @@ class SAM3Backend:
         model = build_sam3_image_model(
             device=builder_device, eval_mode=True,
         )
-        model = model.to(device)
+        model = model.to(device).float()
 
         self.processor = Sam3Processor(
             model,
@@ -260,7 +260,7 @@ class SAM3Backend:
         rgb_array = normalize_slice_to_rgb(slice_2d)
         pil_image = Image.fromarray(rgb_array, mode="RGB")
 
-        with torch.autocast("cuda", dtype=torch.float16):
+        with torch.autocast("cuda", dtype=torch.float32):
             state = self.processor.set_image(pil_image)
             state = self.processor.set_text_prompt(prompt=prompt, state=state)
